# mvcc/kvstore: Optimize compaction