--- a/NOTICE
+++ b/NOTICE
@@ -1,5 +1,0 @@
-CoreOS Project
-Copyright 2014 CoreOS, Inc
-
-This product includes software developed at CoreOS, Inc.
-(http://www.coreos.com/).