@shenjiangc 2019-09-14
> Could you elaborate how this change helps? This is still the same amount of work but is done at a different time? Do you have test data showing the improvement? Thanks!
> 
> Also if we move index compaction to the scheduled FIFO, we need to adjust the metric "indexCompactionPauseMs" accordingly.

Yes, Just let index compact do in another goroutine。

first，I put 1000000 or more key-value：
./tools/benchmark  --target-leader --conns=50 --clients=50 put --key-size=64 --sequential-keys --total=5000000 --val-size=256 --key-space-size=5000000 

then，i do compact use：
rev=$(./etcdctl --endpoints=http://127.0.0.1:2379 endpoint status --write-out="json" | egrep -o '"revision":[0-9]*' | egrep -o '[0-9].*') ; echo $rev
 ./etcdctl --endpoints=http://127.0.0.1:2379 compact $rev
the compaction took more than 1 sec。

when the compact is going on， the other new put request is blocked, and Completion delay > 1 sec.  The reason is the index compact took too time to  assign map[]，when the tree is big enough。

Move index compaction to the scheduled FIFO， let the compaction don't block the new request。

TEST ：（put 5 key，when compacting）
Before Modification：
#./tools/benchmark  --target-leader --conns=1 --clients=1 put --key-size=64 --sequential-keys --total=5 --val-size=256 --key-space-size=5
Summary:
  Total:	2.9459 secs.
  Slowest:	2.9445 secs.
  Fastest:	0.0002 secs.
  Average:	0.5892 secs.
  Stddev:	1.1776 secs.
  Requests/sec:	1.6973
Response time histogram:
  0.0002 [1]	|∎∎∎∎∎∎∎∎∎∎∎∎∎
  0.2947 [3]	|∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎
  0.5891 [0]	|
  0.8835 [0]	|
  1.1779 [0]	|
  1.4723 [0]	|
  1.7668 [0]	|
  2.0612 [0]	|
  2.3556 [0]	|
  2.6500 [0]	|
  2.9445 [1]	|∎∎∎∎∎∎∎∎∎∎∎∎∎

After Modification：
# ./tools/benchmark  --target-leader --conns=1 --clients=1 put --key-size=64 --sequential-keys --total=5 --val-size=256 --key-space-size=5
 Summary:
  Total:	0.0028 secs.
  Slowest:	0.0014 secs.
  Fastest:	0.0003 secs.
  Average:	0.0005 secs.
  Stddev:	0.0004 secs.
  Requests/sec:	1783.4436

Response time histogram:
  0.0003 [1]	|∎∎∎∎∎∎∎∎∎∎∎∎∎
  0.0004 [3]	|∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎
  0.0005 [0]	|
  0.0006 [0]	|
  0.0007 [0]	|
  0.0009 [0]	|
  0.0010 [0]	|
  0.0011 [0]	|
  0.0012 [0]	|
  0.0013 [0]	|
  0.0014 [1]	|∎∎∎∎∎∎∎∎∎∎∎∎∎