@jingyih 2019-09-15
Thanks for doing benchmark testing!

With this change, when the actual index compaction happens later, it will still block other requests such as put request. Is my understanding correct?