@shenjiangc 2019-09-16
With this modification, compaction will no longer block other request messages。;)