@jingyih 2019-09-16
> With this modification, compaction will no longer block other request messages。;)

Sorry I don't quite understand. This modification basically postponed the compaction work in index tree to a later time. So I can understand the compaction request does not block other requests immediately. But later when the actual compaction work happens, it has the same effect on other requests, right?