@shenjiangc 2019-09-16
Hi,
The etch process normal request messages such as put/get/compact in the same goroute “go run()”. When it is processing compact,it will can’t process other messages. The fifo scheduler is another goroute.  So I move the index compact to fifo scheduler, to let the goroute “go run()” to process the next request as soon.