@WIZARD-CXY 2019-10-22
@jingyih I understand your concern. why do we use this as consistency checking?  I think we can allow different etcd node have a little different boltdb state.