@WIZARD-CXY 2019-10-22
@shenjiangc 
>The etcd process normal request messages such as put/get/compact in the same goroute “go run()”. ??
can you show the code?