@jingyih 2019-10-22
> I think we can allow different etcd node have a little different boltdb state.

Do you mean for the same store revision and compact revision, we allow boltdb on each member to have different state? If this is the case, it might be difficult to do consistency checking among members in the same cluster. Today we can use this API to make sure members stay consistent:
https://github.com/etcd-io/etcd/blob/b5afbdd8d0e3098b3d7038fc1e4241de3de494d3/etcdserver/corrupt.go#L353