@WIZARD-CXY 2019-10-23
@jingyih yeah If we enable corrupt check, It may cause the etcd server can not start since peer's hash not match. But in fact I don't enable it...