@jingyih 2019-10-25
> So depends on when the index compaction happens, the returned "keep" could be different, meaning the keys to be deleted in the data base is different.

I think my previous statement was wrong. "keep" does not depend on key modifications after compaction rev.

So this should be safe.