@jingyih 2019-10-25
This conflicts with the tracing we recently added in #11179. 

@YoyinZyc Could you help take a look? If we move the compaction of in-memory index to scheduled, can we still time it? Or maybe we need to drop the corresponding step?

`trace.Step("compact in-memory index tree")`