@jingyih 2019-10-25
Oh actually we also need to adjust the calculation of metrics `indexCompactionPauseMs`.