@YoyinZyc 2019-10-25
> This conflicts with the tracing we recently added in #11179.
> 
> Yuchen Zhou Could you help take a look? If we move the compaction of in-memory index to scheduled, can we still time it? Or maybe we need to drop the corresponding step?
> 
> `trace.Step("compact in-memory index tree")`

We can remove this step in that case. It will be timed with the backend together as a whole scheduled job.