@jingyih 2019-10-26
@shenjiangc Could you please rebase to current master? And then:
1. remove `trace.Step("compact in-memory index tree")`.
2. move the calculation of "indexCompactionPauseMs" to the scheduled function. It is used to measure the duration of `s.kvindex.Compact(rev)`.