@shenjiangc 2019-11-04
> @shenjiangc Could you please rebase to current master? And then:
> 
> 1. remove `trace.Step("compact in-memory index tree")`.
> 2. move the calculation of "indexCompactionPauseMs" to the scheduled function. It is used to measure the duration of `s.kvindex.Compact(rev)`.

I made some mistacks, the new pull request is https://github.com/etcd-io/etcd/pull/11330 , Please take a look。